api/requirements.txt
fastapi
uvicorn
psycopg2-binary
python-dotenv
textblob
pandas
pytest
httpx